+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### 🐞 Bug fixes
 - Fix `queryRenderedFeatures` missing fill-extrusions on globe once it renders as mercator ([#8560](https://github.com/maplibre/maplibre-gl-js/pull/8560)) (by [@birkskyum](https://github.com/birkskyum))
+- Fix `easeTo` and `flyTo` pulling the camera down to the terrain when `centerClampedToGround` is `false`
 - _...Add new stuff here..._
 
 ## 6.11.2